리네임: main의 문자열 계산기코드를 StringCalculator폴더로 이동
